refactor(Conversations): cleanup use of utility functions and typing
