+++ client/src/utils/index.ts
@@ -4,6 +4,7 @@ export * from './latex';
 export * from './presets';
 export * from './languages';
 export * from './endpoints';
+export * from './conversations';
 export { default as cn } from './cn';
 export { default as buildTree } from './buildTree';
 export { default as getLoginError } from './getLoginError';
